refactor(graphics): collapse SetResolution to a single uint32_t overload
Resolution is integer pixels, so keep only SetResolution(uint32_t, uint32_t)
and drop the float and glm::vec2 convenience overloads (the latter unused).
Also remove the unused GetResolution(). The only float caller (PanelViewport,
from an ImVec2) now casts explicitly, making the narrowing visible at the
boundary.

+++ TilesEditor/src/TilesEditor/Panels/PanelViewport.cpp
@@ -48,7 +48,7 @@ namespace Tiles::Editor
         camera->SetSize(m_ViewportSize.x, m_ViewportSize.y);
 
         Tiles::Renderer2D::SetRenderTarget(m_RenderTarget);
-        Tiles::Renderer2D::SetResolution(m_ViewportSize.x, m_ViewportSize.y);
+        Tiles::Renderer2D::SetResolution(static_cast<uint32_t>(m_ViewportSize.x), static_cast<uint32_t>(m_ViewportSize.y));
         Tiles::Renderer2D::Begin(camera);
 
         RenderGrid();
